Update lisk-codec validator logic to better structure code
