+++ elements/lisk-codec/src/index.ts
@@ -13,3 +13,4 @@
  */
 
 export * from './codec';
+export { validateSchema } from './schema';
