+++ elements/lisk-codec/src/schema/errors.ts
