+++ elements/lisk-codec/src/schema/keywords/field_number.ts
@@ -13,7 +13,7 @@
  */
 
 import * as Debug from 'debug';
-import { SchemaError } from '../error';
+import { SchemaError } from '../errors';
 import { Schema } from '../../types';
 
 // eslint-disable-next-line new-cap
